completed video 13: Layers
Added Layers to Hazel.
Encountered problems with runtime errors.  The Cherno talks about some weirdness with the premake file and buildoptions not showing up correctly in VS.

I followed some comments that said to remove the staticruntime="On" commands in the premake file, and that seemed to fix the issue that hew was having in VS.