completed video 15: IMGui
Added support for ImGui, and added a ImGuiLayer to Hazel for displaying things generated by ImGui.

This took a while because the version of the ImGui files have changed a bit since he made the video.  Follow the comments in the YouTube video that say to include 3 files from backend.