essentially done video 22: ImGui Docking and Viewports
Rewrote the ImGuiLayer class to support docking (got rid of a lot of events and we're including some of the ImGui code in the build (see ImGuiBuild.cpp)

Moved ImGuiLayer into Hazel as an overlay (rather having each app need to do it)
Fixed layer stack bug.

There's a challenge at the end of the video that wasn't included in this commit.