part way through video 26 and finally got a static library compile working
the tricky part that caused a lot of link errors was needing to remove --buildoptions from the premake5.lua.  They counteracted the static library setting and essentially changed hazel back to being a dll (I believe).
It showed up in VS as an entry in the ConfigurationProperties->C/C++->CommandLine field.