completed video 42: Textures
After watching the video - my version didn't display correctly (the square was black).  I found a comment that pointed the blame at old computers not running the latest OpenGL (which is true).
That displayed the image - but it was still blurry - so ChatGPT suggested replacing m_RendererID with GL_TEXTURE_2D (in OpenGLTexture.cpp) which seems to be working.