completed video 58: Improving our 2D Rendering API
I did things differently here and opted for a struct for containing default parms (rather than duplicating a bunch of code in functions that do something slightly differently).

I've left the Renderer2D code I wrote with Cherno in but commented it out in case I need to go back in the future.